Make SSL work
If AnyEvent::WebSocket::Client can't find the ssl_ca_file in the default location on your OS this module doesn't work at all. You either need to be able to pass one in, set ssl_no_verify => 1 (which leaves you open to man in the middle attacks), or include a CA, which is what this patch does.